Added in making the venue a link, also radio buttons for 21+
Added in venue being a link(enter url in address input) also added radio
buttons for 21 or over options. Could not get datepicker to work. I
tried using JQuery's datepicker but it wasnt updating the ng-model, so
no values were being added to the collection, even though the date was
showing up in the form. Also sorting of the events is not done by date